@@ -35,13 +35,13 @@
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.1.5",
     "@hookform/resolvers": "^2.7.0",
-    "@material-ui/core": "^5.0.0-alpha.27",
-    "@material-ui/icons": "^5.0.0-alpha.27",
+    "@mui/icons-material": "^5.0.0-rc.0",
+    "@mui/material": "^5.0.0-rc.0",
     "@svgr/webpack": "^5.5.0",
     "@types/formidable": "^1.0.32",
     "@types/googlemaps": "^3.43.2",
     "@types/leaflet": "^1.5.19",
-    "@types/node": "^14.14.10",
+    "@types/node": "^16.7.10",
     "@types/node-fetch": "^2.5.7",
     "@types/nodemailer": "^6.4.0",
     "@types/nprogress": "^0.2.0",
@@ -57,10 +57,10 @@
     "firebase-admin": "^9.4.2",
     "formidable": "^1.2.2",
     "leaflet": "^1.7.1",
-    "next": "^10.0.5",
+    "next": "^11.1.2",
     "next-google-fonts": "^2.2.0",
     "next-sitemap": "^1.6.49",
-    "node-fetch": "^2.6.1",
+    "node-fetch": "^3.0.0",
     "nodemailer": "^6.4.17",
     "nprogress": "^0.2.0",
     "polished": "^4.0.5",
@@ -69,7 +69,7 @@
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.1.0",
     "react-leaflet": "^3.0.4",
-    "sharp": "^0.28.0",
+    "sharp": "^0.29.0",
     "standard-http-error": "^2.0.1",
     "typescript": "^4.1.2",
     "yup": "^0.32.9"
@@ -79,14 +79,15 @@
   },
   "devDependencies": {
     "@emotion/babel-plugin": "^11.2.0",
+    "@next/eslint-plugin-next": "^11.1.2",
     "@typescript-eslint/eslint-plugin": "^4.8.2",
     "@typescript-eslint/parser": "^4.8.2",
     "eslint": "^7.14.0",
     "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "husky": "^6.0.0",
-    "lint-staged": "^10.5.2"
+    "husky": "^7.0.2",
+    "lint-staged": "^11.1.2"
   }
 }
