@@ -2,9 +2,9 @@ import type { Product } from "src/types/model"
 
 import { useState } from "react"
 import styled from "@emotion/styled"
-import { IconButton } from "@material-ui/core"
-import EditIcon from "@material-ui/icons/Edit"
-import DeleteIcon from "@material-ui/icons/Delete"
+import { IconButton } from "@mui/material"
+import EditIcon from "@mui/icons-material/Edit"
+import DeleteIcon from "@mui/icons-material/Delete"
 
 import { Button } from "src/components/Button"
 import Link from "src/components/Link"
