@@ -3,21 +3,21 @@ import styled from "@emotion/styled"
 import NextLink from "next/link"
 import { useRouter } from "next/router"
 
-import List from "@material-ui/core/List"
-import ListItem from "@material-ui/core/ListItem"
-import ListItemText from "@material-ui/core/ListItemText"
-import Divider from "@material-ui/core/Divider"
+import List from "@mui/material/List"
+import ListItem from "@mui/material/ListItem"
+import ListItemText from "@mui/material/ListItemText"
+import Divider from "@mui/material/Divider"
 
-import PersonAddIcon from "@material-ui/icons/PersonAdd"
-import AddCircleIcon from "@material-ui/icons/AddCircle"
-import PersonIcon from "@material-ui/icons/Person"
-import LoginIcon from "@material-ui/icons/Login"
-import LogoutIcon from "@material-ui/icons/Logout"
+import PersonAddIcon from "@mui/icons-material/PersonAdd"
+import AddCircleIcon from "@mui/icons-material/AddCircle"
+import PersonIcon from "@mui/icons-material/Person"
+import LoginIcon from "@mui/icons-material/Login"
+import LogoutIcon from "@mui/icons-material/Logout"
 
 import Loader from "src/components/Loader"
 import { useUser } from "src/helpers/auth"
 import { COLORS, USER_ROLE } from "src/constants"
-import CollectionsIcon from "@material-ui/icons/Collections"
+import CollectionsIcon from "@mui/icons-material/Collections"
 
 const Container = styled.nav`
   width: 300px;
