@@ -11,6 +11,6 @@ export const getFormData = <T = any>(req: NextApiRequest): Promise<[T, SingleFil
       if (err) {
         return reject(err)
       }
-      resolve([(fields as unknown) as T, files as SingleFiles])
+      resolve([fields as unknown as T, files as SingleFiles])
     })
   })
